+++ flutter_app/lib/domains/messages/incoming-messages/MessageObject.dart
@@ -1,4 +1,4 @@
-import 'package:flutterapp/ReceivedMessage.dart';
+import 'package:flutterapp/domains/messages/incoming-messages/ReceivedMessage.dart';
 
 class ObjectMessage extends ReceivedMessage {
   String message;
