+++ flutter_app/lib/domains/messages/incoming-messages/MessageServerTime.dart
@@ -1,4 +1,4 @@
-import 'package:flutterapp/ReceivedMessage.dart';
+import 'package:flutterapp/domains/messages/incoming-messages/ReceivedMessage.dart';
 
 class ServerTimeMessage extends ReceivedMessage{
   int serverTime = 0;
