+++ flutter_app/lib/domains/messages/outgoing-messages/buffering/BufferingContent.dart
