+++ flutter_app/lib/domains/messages/outgoing-messages/chat-message/SendMessageContent.dart
