+++ flutter_app/lib/domains/messages/outgoing-messages/empty-message/EmptyMessage.dart
