+++ flutter_app/lib/domains/messages/outgoing-messages/join-session/JoinSessionContent.dart
