+++ flutter_app/lib/domains/messages/outgoing-messages/join-session/JoinSessionMessage.dart
@@ -1,6 +1,6 @@
-import 'package:flutterapp/domains/messages/join-session/JoinSessionContent.dart';
+import 'package:flutterapp/domains/messages/outgoing-messages/join-session/JoinSessionContent.dart';
 
-import '../Message.dart';
+import '../../Message.dart';
 
 class JoinSessionMessage extends SocketMessage {
   String type = 'joinSession';
