+++ flutter_app/lib/domains/messages/outgoing-messages/join-session/UserSettings.dart
