+++ flutter_app/lib/domains/messages/outgoing-messages/server-time/GetServerTimeContent.dart
