+++ flutter_app/lib/domains/messages/outgoing-messages/update-session/UpdateSessionContent.dart
